Rename NatsCredentialRequest -> AuditTokenRequest
